Sphinx docs: added PDF files as per Fabian's suggestion.
With this change, a PDF form of the user guide and theory book
have been added to the doc/pdf directory (to go along with the other
fairly-static documents).  These need to be put onto the web server
along with the sphinx-generated html files and there is a script
to do this, at least for peterj's login.