Sphinx docs: Have the top-level files pretty much in order.
Have rescued the Theses and Reports content from the old web pages
and found a new home for it in the new order of things.
The actual PDF files for theses are not in the repository but
are in my ~/papers/theses directory and get copied over to the
web server after the Sphinx build is done.