Sphinx docs: Added air-Mach-12.3 example to poshax3 documentation.
In this commit, I have added a poshax3 subdirectory to the sphinx
source area.  I added symlinks (ln -s TARGET LINKNAME) to the
sphinx/poshax3 directory which point to the real files in:
examples/poshax3/air-Mach-12.3.

I also fixed the package names for a Fedora 16 installation
as listed in getting-started.rst