cfpylib: refactored cea2_gas and fixed estcj.py for one of DG's pitot tests.
cea2_gas.py should be a much nicer module now that a lot of redundant code is gone.
Picked up on an occasional error in not applying the power-of-10 multiplier if
there is a space separating the exponent from the fractional part of the printed number.
Also, I hope that I haven't broken too much with the replacement of Shock with shock_process.

The incident shock calculation in estcj now limits the Newton increments to avoid
crazy updates to negative temperatures and the like.
The bad behaviour was detected by David Gildfind back in November 2011.