gas: Changed to 0-based indexing in Lua gas module.
In this commit, I tidied up the nagging issue of
0- vs 1-based indexing when accessing the gas models
through Lua. I've procrastinated on this for a while,
but there was actually minimal code changes required.

Given that there is a user-defined gas model available
in eilmer and that *all* other user input works with
0-based indexing, we decided to change the Lua stuff
to 0-based to be consistent.

The example user-defined gas models and the eilmer3
documentation have been updated accordingly.