cfpylib: First full pass at revising cea2_gas.py.
This revision uses lists to present the CEA2 species data to the user.
It may be better to use a dictionaries: there will be less code and
CEA doesn't seem to order the species the same way as we initially
specify, anyway.  Also, a dictionary for input will be closer to Rowan's
Lua code.