poshax3: Fixed problem with output for air-Mach-12.3 example
I realised that I was plotting mass fractions instead of
moles in my output, and that the details of the kinetic
scheme were incorrect (no ionisation reactions, not
computing reverse rates with the equilibrium constant).

Added options for how species composition is output
in poshax3. Documented new options in poshax3.rst.

Added options to gupta_etal_air_reactions.lua to treat
various number of species. User can select from
5, 7 or 11 (default) species.

Added a version of the Gupta reaction scheme where
the reverse rates are computed using the equilibrium
constants: gupta_etal_with_Keq.lua

Changed the service method in Gas_model which returns
the vector of molecular weights, M(). I made this
return a const reference to the vector. Previously,
it was possible for the caller of the method to alter
the values of M once a reference to the vector had been
passed. This might have lead to unintended consequences.
The new version ensures that only the Gas_model can
change the vector of molecular weights.