estcj: added David Gildfind's pitot-pressure and total-pressure test cases.
Last November David found problems with the ESTCj estimate for pitot pressure
and provided some test cases, one of which showed bad pitot-pressure estimate.
The problem turned out to be a misinterpretation of one of the exponential
number formats used by CEA2.  That was fixed a week or two back.  These tests
should always pass and they're not very long, so I've put them into the main
Eilmer3 test script.