nenzfr: scripts and tests in examples updated to specify nozzle options.
Previously, I had had the Mach 4 nozzle as default in nenzfr and so
didn't specify the nozzle in the test cases based on the Mach 4 nozzle.
Recently, Luke has updated nenzfr to have the Mach 10 nozzle
be the default; now these tests have to explicitly ask for the Mach 4 nozzle.