cfpylib/gasdyn: Added beta_cone2() to take Mach no. as input
Added a second function to compute the shock angle on a cone.
This function takes Mach number as an input. This was a more
natural fit for using when evaluating the Billig correlation
for shock shapes on blunt bodies.

Also, updated the Billig correlation to use the cone angle for
the axisymmetric case. Previously, we'd been using the wedge
angle for all cases when applying the correlation.