eilmer3 example: Manufactured solution test now sets case consistently.
Just for the convenience of setting up this example, we now set the case
by editing the file case.txt.  All of the other scripts should read this
file and set the case accordingly.

Also, we introduce an extra postproicessing script to plot the errors
and we specify the correct gas model file for the postprocessor.