nenzfr: added multi-processor block-marching option.
Initial commit for the changes needed to run nenzfr in a
multi-processor block-marching mode. A new boundary condition
that takes 2 inflow profiles is still needed to make this work
properly. In addition, simulations conducted with nenzfr in
mpi-block-marching mode also needs to be checked against those
in block-sequencing mode to ensure that this new block-marching
option is working correctly.

Another note is that nenzfr to run only with look-up tables
created with ions; it crashes when look-up tables without ions
are used.