Small improvements to cea2_gas and a new builder for LUTs.
For cea2_gas,
(1) added a set_rhoe() method for the look-up-table builder
(2) improved accuracy for very low-pressure conditions,
    first by always using mixture molecular mass to compute
    gas constant and then replacing the read value of p if
    it is small (less than 1kPa).

The look-up-table builder builds on cea2_gas and is thus much
simpler than the older Tcl and Lua codes of the past.
Note that this is the first cut at the new table builder;
it needs finishing so that it can handle gases other than air.

Had to adjust the deltas for some of the unit-test functions
to accommodate the slightly different numbers from cea2_gas.