nenzfr: Working copy of nenzfr in multi-processor block-marching mode
nenzfr can now be run in multi-processor block-marching mode. This
mode can be invoked by adding the --block-marching in nenzfr.py. In
addition, the --nbj option (specifies the number of radial blocks) and
a nozzle.timing file (file that specifies the max_time for each Eilmer3
run) are also needed. See examples/eilmer3/2D/nenzfr-Mach4-nozzle-eq
for an example of how to run nenzfr in this mode.