Eilmer3: Allow a viscous simulation to run when viscosity is zero.
If a gas model uses the "None" mixing rule, transport properties are set to
zero. When used in a viscous simulation, this was causing divide by zero issues
when evaluating the viscous time limit factor of the cell signal frequency.
This factor is now not included if viscosity is zero.