nenzfr_sensitivity: is complete and should work.
This script should be complete. It takes similar inputs to the other
nenzfr scripts except that now the inputs for p1, T1, Vs, pe, Tw,
BLTrans, TurbVisRatio, TurbIntensity and CoreRadiusFraction may be
lists with the first value the nominal  and the other values the
perturbation deltas (as either relative or absolute values).
I am yet to write the script that will read in all the results and
calculate the sensitivity derivatives.

I have also changed the  "Cluster" default to "Mango" for this script
and "nenzfr_quasi_transient.py".