eilmer3: Fixed bug in computing time_slice in main.cxx
In the integrate_blocks_in_sequence function, the time_slice
is currently computed by "G.max_time / G.nblock". Because we
are running 2 blocks in sequence, the number of runs is not
equal to G.nblock, but is equal to (G.nblock - 1). Corrected
this so that we run for the exact max_time that we have
specified.