nenzfr: Automatically creates nozzle.timing file
The nozzle.timing file for the new block-marching mode was
initially created to allow the user more control over the
run-time for each Eilmer3 run in the nenzfr simulation (A
user may want to run for longer durations near the nozzle
throat and shorter durations near the nozzle exit).

Users who do not care about the specific duration for each
Eilmer3 run in nenzfr can omit the nozzle.timing file in the
working directory. nenzfr now automatically creates the file
if it does not exist.