cea2_gas, nenzfr: bug fixes and improvements
For the Mach 4 eq test case with T1=285, cea would return *****e-5 for
the N20 species. This was fixed by changing "trace" to 1.0e-4 in
'make_gas_from_name' for air (within cea2_gas.py). I also found that
the 'clone' function was not copying the "trace" attribute properly.

For nenzfr: The building of eq LUT takes significant time so for
the sensitivity and quasi-transient codes we now build the LUT once,
and then copy it into the case sub-directories. The nozzle supply
temperature and enthalpy have been added to the sensitivity
calculation.

The calculation of the sensitivities is not quite correct.