nenzfr_sensitivity: fixed calc. of sensitivities & uncertainties
The calculation of the sensitivies within "nenzfr_sensitivity_post"
has been checked. We also now calculate and write a summary file of
the uncertainties. These are determined from the sensitivities and
given or default relative uncertainties for each input variable.

I've also cleaned the scripts up a bit and added comments.