nenzfr: Grids are now build such that near-wall cells are wall-orthogonal
For viscous simulations, it is necessary to keep the cells near the non-slip
walls as orthogonal to the walls as possible. However, because the "AO"
option in make_patch does not give a grid that is good enough for the nozzle
geometry, a specialised geometrical surface function is used in the
nozzle.input.template file to build the grids in the expansion region of the
nozzle.