nenzfr_RSA: added ability to calculate residuals
The option '--calculate-residuals' permits the calculation of
residuals for two situations: (1) When creating the Response
Surface, calculate the residuals for each case that was used to
form the RSA. This is a sanity check. (2) When calculating
freestream properties for input (Vs,pe) using the RSA, if a
nenzfr-exit-stats file is available we may also calculate the
residuals.