nenzfr_RSA: added ability to calculate a 2nd-order RSA
When I checked the ability of the radial basis function response
surface to predict freestream properties for the M4 equilibrium
test case, I found that the error was around 2%. This wasn't
acceptable. I have now implemented the ability to use a 2nd order
equation in (Vs, pe) for the response surface. The coefficients
are calculated using least squares regression. This may be
accessed by the --RSA-type option.

I have also added the beginings of a user manual for nenzfr.
What little is there is significantly out of date.