lib/gas: MBWR gas model added.
The Bender gas model can be used for CO2 and H2O. The MBWR gas model
can be used for CO2 and R134a. Note that the two-phase behaviour in
the vapor dome is not taken into account so these models should only
be used in the supercritical, heated vapor and saturated vapor regions.

The list of gas models as been removed from libgas swig interface file
as the avaliable gas models are managed in Rowan's gasfile.lua.