nenzfr: split the parallel and utility functions out into separate files.
The intention is to make the code a bit easier to comprehend,
mainly by reducing the size of files that have to be viewed.
Have also tightened up the interaction with subprocesses so that
commands which go bad are caught.  This gets us to the state that
the original nenzfr (calling e3shared) continues to work but
Wilson's parallel version (calling e3mpi) is clearly broke :(
Sorry Wilson, we'll fix this tomorrow.