nenzfr: Refactored Wilson's parallel procedure to fit my brain.
The principal changes are to the file shuffling, working column-by-column
and naming individual files rather than using glob to grab many files at once.

At this point the parallel version of nenzfr is running in another
window of my workstation; not sure if it will run correctly to completion.
If not, there will be a follow-up commit :)