nenzfr: Removed feature that allows for the control of max run times.
The feature that allows for the control of maximum run times for each
Eilmer3 simulation in the parallel block-marching mode does not get
used at all, and has been removed. The maximum run time for each Eilmer3
simulation is now just the maximum time specified by the user divided
equally by the number of Eilmer3 runs.