cfpylib: added output log to minimize() in nelmin.py
Currently, the minimize() function in nelmin.py is very quiet when
running. For optimisation loops that take a long time to complete,
it is nice for this function to update the user at certain point in
time of its progress. This is done by printing out some statistics
to a nelmin.log file at every optimisation loop.