lib/gas: Added unittest for REFPROP gas model, unit conversion correction.
Fixed mistake in unit conversion for dpdrho in single phase calculation. However,
we cannot use dpdrho for calculating sound speed in a PPF or MIX, must use the
two phase functions. This is because dpdrho in Pa.L/mol when multiplied by the
mixture or pseudo pure fluid molecular weight does not give the right value in
Pa.m^3/kg.