Eilmer3: Make the read_grid method retry if necessary.
This is an attempt to fix the nenzfr problem of failing to find grid files
when running across more than one node on barrine.  We're guessing that the
grid files are a little late in being made available to the newly started
MPI run.

We've also tried to get the MPI run to abort cleanly when things go bad,
however, there is much to be done in passing back appropriate return codes
from functions and actually checking those codes.  Definitely a work
in progress...