Eilmer3: First-order ExtrapolateOutBC, with code that compiles.
In this commit, I did my due diligence and actually checked that
the C++ code compiled before committing.

This commit adds a first-order extrapolation for use with the
extrapolate out boundary condition. It may be accessed from the
Python-level by setting the keyword argument 'x_order=1' when
specifying an ExtrapolateOutBC().