Eilmer3: Extrapolate volumes from interior to ghost cells also.
In addition to the recent change where the positions of ghost
cells are estimated via a linear extrapolation from interior
cells, so too now is the ghost cell volume.

The copying of volume from the nearest interior cell to the ghost
cells for the SUB_OUT boundary condition has now been removed and
replaced with the linear extrapolation. For grids that do not vary
much near a SUB_OUT boundary, this new arrangement for estimating
volume should make very little difference.