eilmer3: Updated udf-bc.lua in MMS case for distorted grids.
The udf-bc.lua file previously assumed that the boundaries
were square located on the edge of the domain:
0 <= x <= L
0 <= y <= L
This is not true for the distorted grid case where the
boundaries are curved.
The interface conditions are now computed from the analytical
solution based on their x,y locations and no assumption is
made with regards to which boundary the interface lies on.