Eilmer3: Fixed e3post.py when using --add-molef option.
Fabian Zander encountered a bug when using the --add-molef
option to e3post.py. Since moving to naming the mass fraction
field as [[ massf[%d]-%s % (isp, specname) ]], that naming
convention was skipped when calling up the mass fraction data
to compute mole fraction. This is fixed in this commit.

HlmerG: Enter commit message.  Lines beginning with 'HG:' are removed.