Eilmer3/examples/2D/sawada_sphere: Updated shock locator.
I changed the trigger pressure for the shock locator
so that it better matches what was previously done
in the locate_shock.awk script. This gives results for
shock detachment distance which better agree with
the non-staged simulation. (Other discrepancies
exist because the grid has been changed for the
staged simulations.)

Also, I increased the simulation lengths to really
be certain that a steady-state was reached.