Eilmer3: Added Gridpro import capability.
Added the ability to import gridpro grids and
property information (boundary conditions) as
service functions callable in the user Python
script.

Also, created a ptymap file which puts hooks
into Gridpro for setting boundary conditions
appropriate for Eilmer. The user of the import
utilities should first install this ptymap
file (ptymap.eilmer) so that the boundary
conditions set in Gridpro are later understood
by Eilmer.

A demonstration of this import capability is
included in:
examples/eilmer3/gridpro-import