Eilmer3/utitilities: Added a simple load-balancing program.
A new utility is added: e3loadbalance.py. It will take
a pre-processed Eilmer job, and an input hosts file
and write out an output hosts file. The input hosts file
should have fewer "processors" (processing units) than the
number of blocks (we'll call these "tasks"). The program
will attempt to find a load-balanced arrangement for assigning
blocks to processors and then write out a new hosts file
for that purpose. The intended use of this program is in
a submission script to a cluster computer. The idea will
be to parse the queue-supplied hosts file
(PBS_NODEFILE, for example), and use the new hosts file
to achieve some degree of load balancing.

This has only been tested for a few of my own contrived
examples. It will get a real kicking tomorrow when I attempt
Rolf's 2308-block simulation.