nenzfr: Changed the way print_stats_CMME() computes the 1D flow properties.
Rather than follow Baurle and Gaffney's paper, we just minimize the errors in
the conserved properties directly with the nelmin optimizer.  This brute-force
approach has the significant advantage that it works for any gas model and
has the added advantage for me (PJ) that it is lots easier to understand.