Eilmer3: Added a test directory for the load-balancing work.
This test directory contains a Gridpro grid with 27 blocks.
In our demo, we request a distribution of those 27 blocks
over 16 processors.

Also, made e3loadbalance.py sort the blocks by id before
listing them in the mpimap file. I don't think this order
is important but it was nice to scan them in order
from a human perspective. In reality, the block counts
might be so large that a human wouldn't really want to
look at this file anyway.