Eilmer3: basic code for handling a bag of blocks for each MPI task.
This is stage 1 of several stages of getting the MPI version of the code
to be able to handle several blocks per MPI task.
Each process keeps a vector of pointers to the blocks for which it is
responsible.  No other changes have been made.
This code compiles but has not been tested. Do not use this commit.