Eilmer3: Updated visc3D.cxx.m4 to be consistent with visc3D.cxx
Yes, that top line is correct. visc3D.cxx.m4 was out-of-date
w.r.t. visc3D.cxx. (If this happens, the normal assumption would
be that it would be the other way around... visc3D.cxx is out-of-sync
with visc3D.cxx.m4.) The calls to modal_enthalpy evaluation were
not up-to-date with the method signature in lib/gas/models/gas-model.hh.