NENZFr: Added example to showcase the optimisation of a T4 nozzle.
This example demonstrates the optimisation of the Mach 7 nozzle by
coupling the in-built Nelder & Mead optimiser with NENZFr. This is
the first commit of this example, so it may contain some minor bugs.
We will attempt to weed these bugs out as Han Wei attempts to use
these scripts over the next couple of weeks.