Eilmer3: Some debug printing added for Rolf and Rowan's gridpro duct.
This debug has been hidden from the compiler with #if 0 .. #endif macros
so that it should not interfere with normal code operation.
In cell.cxx, we now print the area and unit vectors associated with
the cell faces.  Part of our troubles appears to be a lot of sensitivity
of the values of these vectors to the cell vertex positions that is
fed through to the flow solution.  Well, it looks that way, presently.
Also shifted uflowz from e3_flow.py to e3_grid.py so that it could be
used in both modules.  e3_flow imports e3_grid already.