Eilmer3: Introduction of house-keeping code for MHD physics.
This is mainly adding the magnetic field vector B to the flow data
and mhd_flag to the kernel to allow the MHD physics code (yet to be
added by Daryl Bond) to the activated.

Also, a couple of small fixes to the LaTeX files for the user guide
so that it would build cleanly.