Eilmer3: postprocessor now writes a .pvd collection file for Paraview.
This file is useful for picking up a time-series of solutions.

Also, there are a few new comments written into the Tecplot ASCII-format
data file to make it a little easier to browse.
Although this was done in an attempt to fix Rolf's Tecplot plotting issue
at block boundaries, no actual code bug could be identified.