lib/gas/tools: When looking for species search for $E3BIN first, then $HOME/e3bin
In this commit, I've relaxed the hard-coded assumption that files will always
be placed in $HOME/e3bin. The tools will now look for a user-set evironment
variable E3BIN which points to the e3bin directory. If not set, it will
default back to looking in $HOME/e3bin.

[Spotted a large amount of technical debt in the lib/gas/tools code.
There is a *lot* of duplicated code in each of the lua modules.]