lib/gas: Warnings printed if missing atomic_constituents or charge.
When creating a thermally_perfect_gas model, warnings are now issued
is the requested species are missing information about atomic_constituents
or charge. So that the building of a gas model can proceed, it will supply
some default values found in species/default.lua.

TODO: Have only fixed the script that makes a thermally_perfect_gas and
not all of the other composite gas types. What really needs to happen
is that the script to create a composite gas type needs to be unified.
(Try doing that this week.)