examples/eilmer3: continued work on Hayabusa forebody example
The example will (eventually) contain 4 parts:

1)  inviscid simulation on a coarse mesh
1b) inviscid simulation on a coarse mesh with radiation coupling
2)  viscous simulation on a refined mesh
3)  viscous simulation on a refined mesh with radiation coupling

Part 1b) is included so as to provide an example of radiation coupling
that runs in a relatively short amount of time.

Notable modifications in this commit include the use of the
ShockLayerSurface class to ensure cells adjacent to the wall are
normal to the surface.

Peter - at the moment I have put this class in:

lib/cfpylib/flow/shock_layer_surface.py

It operates in a similar way to make_patch().  Is there a better place
to store it?