poshax3 and radiation2: a scaling factor for div(qrad) when applied to electronic energy equation
When we are dealing with partially electronic level specific chemistry
(i.e. some species are electronic levels, others are bulk species),
the energy of radiating levels may or may not be part of the electronic
energy, depending on if a Boltzmann distribution is assumed or not.
This factor allows this effect to be approximated.