Eilmer3 examples: added 3D mms test case.
This is the first cut at bringing across the 2D test case.
Already, for case 1, we see that there is a small nonuniformity
in the z-direction where there really shouldn't be any.
This nonuniformity exists irrespective of the x-orders of
the main reconstruction and the extrapolate-out boundary conditions
on the north and east boundaries.