Encoded a hyperbolic tangent clustering function in the c++ code.
This one is from http://www.cfd-online.com/Wiki/Structured_mesh_generation
and differs from Adrian's version in that it only requires the clustering
gradients at either end as input, instead of these plus a path length and
discretisation, which is more suitable for how the cluster functions are
used in defining eilmer3 meshes.

With this installed, it is now possible to make shock layer grids
with different clustering at the boundary layer and and shock front using a
DiscontinuousUnivariateFunction.  Will upload an examle soon.