Eilmer3/sphinx: Fixed error in example .bashrc file for working on barrine.
Han identified an error related to using the module command.
The instructions I had up were to include a line:
  module load purge
when it really should be:
  module purge

As a side note, the purging of modules at the start of a session
seems to be necessary to get rid of some default loaded
modules which interfere with the intel-mpi environment. When asking
David Green about this problem, his recommendation was to
begin a session with 'module purge'.