Eilmer3: Example to test PJ's new technique of calculating wall derivatives
Added an example to aid PJ test his implementation of the improved
technique to compute wall derivatives. This test case computes the
interaction of a shock wave generated by a 20-deg wedge with a
turbulent boundary layer. The results show very strong (and
unphysical) checkerboard patterns on the surface of the 20-deg
wedge. Hopefully, PJ's new technique will resolve this checkerboard
issue.