examples/eilmer3/2D/hayabusa/part2-viscous/hayabusa.py:
Implemented an inflow boundary that is fitted to the shape of the shock.
The shock locus is determined by finding the flow data points with S=1
and using the scipy.optimise toolbox to fit a Billig curve to the point
cloud.  The inflow boundary is then determined by extending the normal distance
between the body and shock loci by a constant factor (1.2 in this case).

Also the grid is now clustered at both the shock and body thanks to the
HypertanClusterFunction used in conjunction with a RobertsClusterFunction
as a DiscontinuousUnivariateFunction.

A number of things need improving and checking:
 - a Billig curve is not the best way to describe the shock locus, even with all the parameters open to
   the optimisation procedure (and x and y scaling factors).
   Something such as a spline could be used, however care must be taken to ensure dy/dx = infinity at y=0
   so that the grid is nicely formed at the stagnation point (this is why a billig curve was used for this initial implementation).
 - the cell size against the wall needs to be checked to ensure it is similar to the previous version where
   roberts clustering with beta=1.01 was used