lib/gas/tools: include the species charge for the thermal nonequilibrium gas-models
Previously this was not being included, but for some reason only became a problem
when attempting to run on barrine with the intel compiler.

Also it should be noted that the charge for many species has previously been included in the library
under the 'Z' field.  We need to decide whether to proceed with 'Z' or 'charge' as
otherwise we are doubling up on this information.