lib/gas: Fixed bug related to looking for E3BIN
In this commit, the kinetics module will look
for the existence of the environment variable
$E3BIN. From this it will derive the location
of the Lua parsers for reactions and energy
exchange mechanisms. If $E3BIN is not set, then
the code will default to looking in $HOME/e3bin.
So there should be no change in behaviour for users
who install the code in the default location.