lib/gas: Bug fix in VT-Millikan-White related to parsing input.
When picking up one of the parameters for the Millikan-White
VT relaxation time model, the lua stack was not being left
as it was found. I had forgotten an extra lua_pop(). This
problem only became apparent when there were multiple
relaxation times for the one mode AND the Millikan-White
was listed early in the list of rates.