onedval: Initial commit of a work in progress.
Onedval is a small utility which can work on a slice of
CFD data to give one-dimensionalised quantities. The
one-dimensionalisation techniques borrow from ideas in
nenzfr. Onedval will work on any 2D data slice extracted
with Tecplot.

In the present commit, all of the number-crunching
one-dimensionalizing techniques are built.

TODO: Prettify the user interaction.