lib/gas: beginnings of 'Fully_coupled_polyatom_internal' species energy mode.
Currently only the partition function for a linear triatomic species can be calculated.
This class will be used as a development area to get the thermodynamics correct for
polyatomic species where the internal modes consider some degree of coupling.
Eventually these functions can also be transfered over to the radiation module
to allow polyatomic radiation to be calculated.