app/onedval: Fixed problem with orientation of cell normals.
The finite-element data produced by Tecplot does NOT
list the corners of cells in a manner that produces
a consisten orientation of cell normals. I had situations
where adjacent cells would switch between an
outward-facing normal, and in inward-facing normal.
This switching played havoc with computing the fluxes
across the interface.

The fix was to hard-code an assumption about the x-direction
of the cell normals to give a consistent orientation.
This hack will work reliably for slices of 2D data across the yz plane
which is presently the largest use case for onedval.
A more cunning long-term solution would involve the construction
of my own 2D surface from nodal data.