lib/gas/species/HCO.lua and HCN.lua - added/updated species files.
Non-equilibrium data was added for HCN, and HCO was created, as both
of these species are required for air + ablation chemistry models.
With the inclusion of these files, all species are present to run the
three most common models - Park, Abe and Olynick. More work to come
on reactions and collision integrals.