Eilmer3: Added a function to check parameter constraints when preparing a job.
This function needs to be filled out with logic that checks and applies
constraints to the parameters that are about to be written to the configuration
and control files.  This will happen as we think of things.
The first item is a check that sets the moving-grid flag if the shock-fitting
flag has been set.