Eilmer3 automated tests: added a simple Sod shock tube case.
This uses the simple shock tube to test that the low-temperature
thermally-perfect gas mix acts just like the ideal-gas mix
without giving any grief over negative values for internal energies.
Revision 820 prompted the inclusion of this test; it had been on
the agenda for years, to make sure that we had weeded out all of
the ideal-gas assumptions that might give trouble when a non-ideal
gas (which does not have e=Cv*T) was used.