Eilmer3 house-keeping: separate fixed-grid and moving-grid convective updates.
Essentially copy Andrew's intermixed-update code to a new fixed-grid procedure
and strip out the moving-grid aspects.
The short automated tests are working fine (so far up to the sawada sphere) and
are working as before, so the next step is to propagate Andrew's vector quantities
to other parts of the code and merge the essentially duplicated functions that
he had to introduce a couple of months back.