lib/util: Added a new TARGET to systems.mk -- for_gnu_local
In this commit, a new compilation TARGET for_gnu_local has been
added. This allows user with a local, non-default install of the
GNU compiler suite to build the code. The idea is that the user
might have a local install of the GNU compilers with a suffix added
to the compiler names. For example, on my machine I recently installed
GNU GCC 4.8.0 into /usr/local with a program suffix of "-4.8". To
use this compiler with the new TARGET, I issue the following:

> make TARGET=for_gnu_local GNU_SUFFIX=4.8