Eilmer3 house-keeping: start transition from fixed-grid scalars to moving-grid vectors.
At this point, Andrew Pastrello's vector quantities have been propagated through
the code and the build is clean (using gcc 4.8.0), however, the fixed-grid
formulation is completely untested and the moving-grid behaviour is known to be
broken.  I could say that this commit is a preview for Andrew but the reality is
that I want to go to the pub with the rest of the fellows.

One nice thing to come out of the work is the elimination of a bug in the 3D
viscous derivative code.  The change is marked with a comment in visc3D.cxx.m4.
Maybe this will help with Wilson's turbulent wedge calculations.