Eilmer3 house-keeping: separated the moving-grid and fixed-grid flow-update functions.
This allows the fixed-grid flow-update to return to a much simpler form and allows
the moving-grid flow-update functions to have a simpler implementation that doesn't
have to worry about too many configuration flags.
The cost is some nearly-duplicated code that has to be maintained but, at least,
the simpler forms of the update functions are less likely to make my head explode.