Eilmer3 house-keeping: clean-up the selection of the gasdynamic update scheme.
The notation for selecting and applying the gasdynamic time-stepping scheme for
fixed-grid simulations is improved and made a bit more versatile.  Euler and
predictor-corrector schemes should work as before.  It should be easy to
implement further options cleanly.

A new control parameter is introduced, gdata.gasdynamic_update_scheme, that
take a string value ("euler", "pc", "predictor-corrector").
It replaces (the now deprecated) gdata.t_order, however, gdata.t_order will
continue to work so that older simulations should not break because of this
particular change :)