Eilmer3 house-keeping: replaced U_old and U in the FV_Cell with a std:vector.
This generalization will allow us to implement a variety of flow update schemes
more clearly and conveniently.  Although this revision builds and runs the cone20
test case, it is for review as rather than serious use.  More work is to come in
both testing viscous cases and in implementing some new update schemes.