Eilmer3 house-keeping: reinstate the viscous update within the gasdynamic update.
The default behaviour for viscous simulations with explicit time-stepping is, once
again, to include the viscous terms with the inviscid update.  This means that the
explicit viscous updates get the benefit of the full selection of update schemes
and, hopefully, will be a bit better behaved than the separate Euler update for
viscous terms.

The option for doing the viscous increment separately is still available as an
option by setting gdata.separate_update_viscous_flag = 1 in you input script.

Also, setting the flag for implicit updates will still do the viscous increment
separately from the implicit-inviscid update, however, the implicit-update code
is not functional at the moment.

Along the way, we had to more carefully label source-vector additions as being
such.  Source vector quantities are now cleared to zero at the start of each
stage of the gasdynamic update and are then accumulated at each call to a source
vector service function (for inviscid, viscous and user-defined).