onedval: Increased robustness by always using internal gas model to compute energy flux.
Some difficulties were encountered when processing some data
provided by Mat Bricalli. In particular, the optimiser was failing
to converge (and giving very poor results) in regions where
fuel was first injected in the flow. It turns out the problem was related
to a mismatch between my internal gas model (thermo curves based
on latest CEA curves, and choice of reference energy) compared to the
thermodynamic model used in CFD++. CFD++ uses the old form of NASA
curve fits for the thermo evaluations and it's not clear what reference
point is used for energy or enthalpy. The outcome was that my model would not
compute the same stagnation enthalpy at the same conditions, so the optimiser
was struggling to find a temperature value which would satisfy the energy
flux requirement.

The solution is that onedval only uses temperature (which does not suffer from
any arbitrary reference point) to reconstruct the enthalpy and from that
deduce the energy flux across the slice. The same internal gas model is then
used when trying to find a set of flux conserved values. The calculation is now a little
slower because the interal gas model is called for every cell instead of relying
on the CFD-supplied h0 value. However, using a consistent gas model has seemed
to eliminate any robustness problems.

Now that the calculation is more robust, the hack I introduced by using a brute force
optimiser has been removed. This cleans up some of the code and removes what I
thought was a bad hack anyway, so I'm pleased about this.

Ths user guide has been updated to drop the stagnation/total enthalpy requirement
and instead replace it with a temperature requirement.