lib/gas; poshax: Changed energy vector.
In this commit, I changed the convention of what gets stored
in the Gas_data e[] vector. Previously, e[0] had the special
status of storing the TOTAL internal energy of the gas. In
any internal thermo routines, we often had to recompute the
energy in just the translational mode.

Now, e[0] has no special status. It will typically store
the transrotational energy, but the form of what is stored
is no different to any of the other components.